+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## v0.15
 
+* `bayes_net` can now be cached to a file. Use `dump_cache_file` to write the file and `load_cache_file` to load from the file.
 * Distributions can now be compared with `==` and `!=`.